# [crypto] Faster modular multiplication for P256