--- a/sw/otbn/crypto/p256_shared_key.s
+++ b/sw/otbn/crypto/p256_shared_key.s
@@ -111,7 +111,7 @@ p256_shared_key:
 
   /* Compute modular multiplication of m_x and z^-1.
      w19 = w24 * w25 mod p = m_x * z^-1 mod p = x1 */
-  jal       x1, mod_mul_256x256
+  jal       x1, mul_modp
 
   /* Store "affine" mask to DMEM. Use the y-coordinate
      to save memory (not needed afterwards)