… sw/otbn/crypto/p256_base.s
… sw/otbn/crypto/tests/p256_mul_modp_test.exp