This change implements a faster modular reduction algorithm that is specialized to the modulus `p` of the coordinate field for P256. (The older Barrett reduction is still used for the scalar field modulo the curve order `n`, so this costs a bit of code size.) 

The new modular multiplication is 52 instructions, compared to 65 for the Barrett-based version. The number of instructions should match the number of cycles in this case, since there are no loads/stores, so this is about a 1.25x speedup in the core routine. The coordinate field is pretty performance-critical, so it translates to a 1.23x speedup in top-level ECDSA sign/verify performance:

| operation | master | this PR | speedup |
|----------------|------------|-----------|---------------|
| sign           | 867291 | 704126 | 1.23x      |
| verify         | 517460 | 420220 | 1.23x     |


## Proof

Because the new algorithm is a bit mathematically complex, I verified it in Coq to make sure I wasn't missing any corner cases: https://gist.github.com/jadephilipoom/5c1910fd355f730238c99ce620aed98a (the link is also in the docstring)

Read on for some guidance if you'd like to review the proof. You should only need to review the final proof statement and its sub-definitions like `exec`; the bodies of proofs are just hints to Coq about how to solve the proof and should be OK as long as Coq can typecheck them.

That Coq file defines a simplified OTBN model that only includes the instructions/functionality that this code uses, and then some notations that made it possible to copy-paste the code. Then it has quite a bit of math and boilerplate, but the punchline is the final proof, which says:
```coq
  Lemma femul_func_correct regf flagf a b :
    0 <= a < p ->
    0 <= b < p ->
    has_reg regf w24 a ->
    has_reg regf w25 b ->
    has_reg regf w28 r256 ->
    has_reg regf w29 r448 ->
    has_reg regf w31 0 ->
    has_reg regf MOD p ->
    exec regf flagf femul_func
         (fun regf' _ =>
            clobbered regf regf' [(w19 : reg)
                                    ; (w20 : reg)
                                    ; (w21 : reg)
                                    ; (w22 : reg)
                                    ; (w23 : reg)
                                    ; (w24 : reg)
                                    ; (w25 : reg)
                                    ; (ACC : reg)]
            /\ get_reg regf' w19 = (a * b) mod p).
```
Here, `femul_func` is the OTBN code, which I copy-pasted from the implementation. That means that for any starting register file `regf` and flag file `flagf`, and any integers `a` and `b`, if `regf` contains the entries:
- `w24=a`
- `w25=b`
- `w28=r256` (`r256` is a constant value defined earlier, `r256 = 2^256 mod p`)
- `w29=r448` (`r448` is also a constant, `r448 = 2^448 mod p`)
- `w31=0`
- `MOD=p`
...and `a` and `b` are reduced mod p, then executing `femul_func` with the OTBN model over the code (`exec regf flagf femul_func`) starting from that register/flag state produces a new state with register file state `regf'` such that the only registers clobbered between `regf` and `regf'` are `w19` to `w25` and `ACC`, and register `w19` holds the value `(a * b) mod p`.