@jadephilipoom 2023-12-22
> w00t! Thanks @jadephilipoom!
> 
> Thanks for the details added to the PR description. I wonder if it would make sense to add a section covering implementation notes to the general documentation. It has been often the case that people look for this level of detail from time to time when digesting the code.
> 
> Alternatively, you could add a reference to this PR in the `mul_modp` stringdoc.

For now I just added the link to this PR to the docstring. I think documentation pages for some of the cryptolib implementation stuff is a good idea, though, so I created https://github.com/lowRISC/opentitan/issues/20723 to track this for at least P256 modmul and a writeup about RSA keygen that I've already done but not put anywhere on the website yet.